# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Web service for comic book data — tiered API access via API key.\n\n## Authentication\nAll `/api/v1/*` endpoints require the `X-API-Key` header.\nObtain an API key via `POST /auth/register` or `POST /auth/api-key`.\n\n## Plans\n| Plan | Quota | Rate Limit | Features |\n|---|---|---|---|\n| free | 1,000 req/month | 10 req/min | list, detail, chapters |\n| standard | 100,000 req/month | 120 req/min | + search |\n| premium | unlimited | 1,000 req/min | + recommend, analytics |",
        "title": "Comic Provider API",
        "contact": {
            "name": "Comic Provider Support"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/comics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a paginated list of comics. Available to all plans.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comics"
                ],
                "summary": "List comics",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of results per page (1-100, default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data array of comics",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid_query",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "rate_limit_exceeded / quota_exceeded",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/comics/recommend": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns comics similar to the given comic ID or title. Requires premium plan.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comics"
                ],
                "summary": "Recommend similar comics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base comic ID to find similar comics",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Base comic title to find similar comics",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (1-50, default 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data array of recommended comics",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid_query (id or title required)",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "feature_not_allowed (non-premium)",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "comic_not_found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "rate_limit_exceeded / quota_exceeded",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/comics/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Search comics by various filters. Standard plan supports title/genre/author/country. Premium plan additionally supports age_rating, type, sort, order, year, year_from, year_to.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comics"
                ],
                "summary": "Search comics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Keyword search by title (alias: title)",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Title keyword (alias of q)",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Genre/category slug (alias: category)",
                        "name": "genre",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category slug",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Author name keyword",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Country keyword",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Age rating filter (premium only)",
                        "name": "age_rating",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Book type: manga, manhua, manhwa, comic, lightnovel (premium only)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field: created_at, publish_year, title (premium only)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order: asc, desc (premium only)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Exact publish year (premium only)",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Publish year from (premium only)",
                        "name": "year_from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Publish year to (premium only)",
                        "name": "year_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Results per page (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data array of comics",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid_query",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "feature_not_allowed (free plan)",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "rate_limit_exceeded / quota_exceeded",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/comics/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns detailed information for a single comic by ID. Available to all plans.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comics"
                ],
                "summary": "Get comic detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data object of comic",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "comic_not_found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "rate_limit_exceeded / quota_exceeded",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/comics/{id}/chapters": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all chapters for the given comic ID. Available to all plans.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comics"
                ],
                "summary": "List chapters of a comic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data array of chapters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "comic_not_found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "rate_limit_exceeded / quota_exceeded",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for accessing protected /api/v1/* endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}